// app copy/container/Tariff/interface.ts
export interface PromotionalSummaryCardOffer {
    /**
     * Offer Price
     */
    offerPrice: number | null;
}

/**
 * Initial state
 */
export interface IInitialState {
    /**
     * Loading flags per API
     */
    loading: {
        getSubscription: boolean;
    };
    /**
     * Error flags per API
     */
    errors: {
        getSubscription: boolean;
    };
    /**
     * Subscription Id
     */
    subscriptionId: RedTariff | YoungTariff | null;
    /**
     * offers
     */
    subscriptionPayload: TariffWithHardwareResponse | null;
    /**
     * promotional Summary Card Offer
     */
    promotionalSummaryCardOffer: PromotionalSummaryCardOffer,
}